requirements.txt
livekit-api
fastapi
uvicorn
pydantic
python-dotenv~=1.0